Fix rename_dir() when target already exists
shutil.move() used in rename_dir() does not raise FileExistsErroro if
the target is an existing directory, contrary to what the code expected.
Instead, shutil.move() moves the destination inside the target
directory. This is not a renaming operation as the function's name
suggests. We now check the existence of target directory before calling
shutil.move() and bail out if it exists.

Re #620

+++ spinetoolbox/helpers.py
@@ -328,6 +328,13 @@ def rename_dir(old_dir, new_dir, logger):
         new_dir (str): Absolute path to new directory
         logger (LoggerInterface): A logger instance
     """
+    if os.path.exists(new_dir):
+        # If the target is a directory, then there will not be a name clash in shutil.move()
+        # as the old_dir will be moved inside new_dir which is not a rename operation.
+        # We guard against that here.
+        msg = "Directory<br/><b>{0}</b><br/>already exists".format(new_dir)
+        logger.information_box.emit("Renaming directory failed", msg)
+        return False
     try:
         shutil.move(old_dir, new_dir)
     except FileExistsError:
